Module salvo_extra::auth::jwt

Structs

CookieExtractor

DecodingKey
All the different kind of keys we can use to decode a JWT This key can be re-used so make sure you only initialize it once if you can for better performance

FormExtractor

HeaderExtractor

JwtConfig

JwtError
An error that can occur when encoding/decoding JWTs

JwtHandler

QueryExtractor

TokenData
The return type of a successful call to decode.

Validation
Contains the various validations that are applied after decoding a JWT.

Enums

Algorithm
The algorithms supported for signing/verifying JWTs

Traits

JwtExtractor

Functions

decode
Decode and validate a JWT