Module salvo_extra::auth::jwt

Structs

CookieExtractor

FormExtractor

HeaderExtractor

JwtConfig

JwtError
An error that can occur when encoding/decoding JWTs

JwtHandler

QueryExtractor

TokenData
The return type of a successful call to decode

Validation
Contains the various validations that are applied after decoding a token.

Enums

Algorithm
The algorithms supported for signing/verifying

Traits

JwtExtractor

Functions

decode
Decode a token into a struct containing 2 fields: claims and header.